dnsdist: Move CacheValue definition into a standalone header
`cxx::bridge` needs to be able to parse the header that has the
definition of `CacheValue`. The `*.rs.h` header files generated by
`cxx::bridge` won't be available until after `cxx::bridge` has finished
building (i.e., until after dnsdist-rust-lib has been built).

That means that any header file that `cxx::bridge` wants to read via
an `include!` directive cannot itself have a dependency on any of the
header files generated by `cxx::bridge`, because those generated header
files won't actually exist (if this is a clean working tree).

Since dnsdist-cache.hh has an include of `rust/moka.rs.h`, this causes
a build failure when `cxx::bridge` runs and can't find that header file.
So, this commit breaks out the `CacheValue` struct into its own header
file that doesn't have a dependency on any of the headers generated
by `cxx::bridge`.
